+++ newCluster.h
@@ -6,7 +6,7 @@
 #ifndef NEWCLUSTER_H
 #define NEWCLUSTER_H
 
-int clusterOptimum(NetArch netA, NodeArch nodeA, float dBS);
+int clusterOptimum(NetArch netA, NodeArch nodeA, double dBS);
 
 ClusterModel newCluster(NetArch netA, NodeArch nodeA, int roundNo, int p_numCluster);
